% Generated by roxygen2 (4.0.0.99): do not edit by hand
\name{multi_plot}
\alias{multi_plot}
\title{multi_plot}
\usage{
multi_plot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored}

\item{plotlist}{A list of plots as opposed to the series}
}
\description{
ggplot objects can be passed into this function to create multiple plots on one page. Created by www.cookbook-r.com
}
\examples{
multiplot(p1, p2, p3, p4, cols=2)
}
\keyword{ggplot2}
\keyword{multiple}
\keyword{plot}

