% Generated by roxygen2 (4.0.0.99): do not edit by hand
\docType{package}
\name{plot_res}
\alias{plot_res}
\alias{plot_res-package}
\title{plot_res.}
\usage{
plot_res(mod)
}
\arguments{
\item{base_size}{sets the base font size for graphs - text sizes are proportional to this value}
}
\description{
plot_res.

A simple plot of residuals
}
\examples{
plot_res(model)
}
\keyword{ggplot2}
\keyword{models}
\keyword{residuals}

