% Generated by roxygen2 (4.0.0.99): do not edit by hand
\docType{package}
\name{theme_evp}
\alias{theme_evp}
\alias{theme_evp-package}
\title{theme_evp.}
\usage{
theme_evp(base_size = 18, base_family = "")
}
\arguments{
\item{base_size}{sets the base font size for graphs - text sizes are proportional to this value}
}
\description{
theme_evp.

A theme for ggplot created for Evan's favourite options
}
\examples{
ggplot(data, aes(x=x,y=y)) +
 geom_point() +
 theme_evp()
}
\keyword{ggplot2}
\keyword{theme}

